青福换电

+++ src/main/java/com/xiliulou/electricity/mapper/ElectricityMemberCardMapper.java
@@ -16,7 +16,7 @@ public interface ElectricityMemberCardMapper extends BaseMapper<ElectricityMembe
 
 	Integer queryCount(@Param("status") Integer status, @Param("type") Integer type,@Param("tenantId") Integer tenantId);
 
-	List<ElectricityMemberCard> listByFranchisee(@Param("offset") Long offset, @Param("size") Long size, @Param("status") Integer status, @Param("type") Integer type,@Param("tenantId") Integer tenantId, @Param("franchiseeId") Integer franchiseeId);
+	List<ElectricityMemberCard> listByFranchisee(@Param("offset") Long offset, @Param("size") Long size, @Param("status") Integer status, @Param("type") Integer type,@Param("tenantId") Integer tenantId, @Param("franchiseeId") Long franchiseeId);
 
-	Integer listCountByFranchisee(@Param("status") Integer status, @Param("type") Integer type,@Param("tenantId") Integer tenantId, @Param("franchiseeId") Integer franchiseeId);
+	Integer listCountByFranchisee(@Param("status") Integer status, @Param("type") Integer type,@Param("tenantId") Integer tenantId, @Param("franchiseeId") Long franchiseeId);
 }
